requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
diffusers==0.2.4
filelock==3.8.0
h11==0.13.0
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
tokenizers==0.12.1
tqdm==4.64.0
transformers==4.21.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
Werkzeug==2.2.2
zipp==3.8.1
anyio==3.6.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
diffusers==0.2.4
fastapi==0.81.0
filelock==3.8.0
Flask==2.2.2
h11==0.13.0
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pydantic==1.9.2
pyparsing==3.0.9
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
sniffio==1.2.0
starlette==0.19.1
tokenizers==0.12.1
tqdm==4.64.0
transformers==4.21.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
Werkzeug==2.2.2
zipp==3.8.1
